# docs/openapi.yaml
openapi: 3.0.0
info:
  title: PROYECTO BASE BACKEND
  version: '1.0'
  description: Documentación del Proyecto base BACKEND
  contact:
    name: agetic@agetic.gob.bo
servers:
  - url: 'http://olimpiadas-backend.test.gtic.gob.bo/olimpiadas-backend'  
paths:
...
  /api/usuarios/activacion/{id}:
    patch:
      summary: Activa un usuario
      tags: ['usuario']
      operationId: get-api-v1-usuarios-activacion
      description: Api para activar un usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            example: 'a9d1a5cc-4590-5c67-a0b2-a4b37b862802'
          description: id de usuario
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje: 
                    type: string
                    example: 'Registro actualizado con exito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                      estado:
                        type: string
                        example: 'ACTIVO'
...
  /api/usuarios:
    get:
      summary: Api para obtener el listado de usuarios
      tags: ['usuario']
      operationId: get-api-usuarios
      description: Api para obtener el listado de usuarios
      parameters:
        - in: query
          name: pagina
          schema:
            type: string
            example: 1
          description: numero de la pagina a visualizar
        - in: query
          name: limite
          schema:
            type: string
            example: 10
          description: numero de filas a visualizar
        - in: query
          name: orden
          schema:
            type: string
            example: ASC
          description: tipo de orden a visualizar
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje: 
                    type: string
                    example: 'Registro(s) obtenido(s) con exito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      filas:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                            usuario:
                              type: string
                              example: 'ADMINISTRADOR'
                            estado:  
                              type: string
                              example: 'ACTIVO'
                            persona:
                              type: object
                              properties:
                                nroDocumento:
                                  type: string
                                  example: '1234567'
                                nombres:
                                  type: string
                                  example: 'JUAN'
                                primerApellido:
                                  type: string
                                  example: 'PEREZ'
                                segundoApellido:
                                  type: string
                                  example: 'PEREZ'
                            usuarioRol:
                              type: array
                              items:
                                type: object
                                properties:
                                  estado:
                                    type: string
                                    example: 'ACTIVO'
                                  id:
                                    type: string
                                    example: 'd5de12df-3cc3-5a58-a742-be24030482d8'
                                  rol:
                                    type: object
                                    properties:
                                      rol: 
                                        type: string
                                        example: 'ADMINISTRADOR'
                      total:
                        type: number
                        example: 1
    post:
      summary: Api para crear una nuevo usuario
      tags: ['usuario']
      operationId: post-api-usuarios
      description: Api para crear un nuevo usuario
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                usuario:
                  type: string
                  example: '123456'
                contrasena:
                  type: string
                  example: 'password'
                persona:
                  type: object
                  properties:
                    nombres:
                      type: string
                      example: 'JUAN'
                    primerApellido:
                      type: string
                      example: 'PEREZ'
                    segundoApellido:
                      type: string
                      example: 'PEREZ'
                    nroDocumento:
                      type: string
                      example: '123456'
                    fechaNacimiento:
                      type: string
                      example: '1999-11-11'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: 'Registro creado con exito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      usuario:
                        type: string
                        example: '123456'
                      estado:
                        type: string
                        example: 'CREADO'
                      persona:
                        type: object
                        properties:
                          nombres:
                            type: string
                            example: 'JUAN'
                          primerApellido:
                            type: string
                            example: 'PEREZ'
                          segundoApellido:
                            type: string
                            example: 'PEREZ'
                          nroDocumento:
                            type: string
                            example: '123456'
                          fechaNacimiento:
                            type: string
                            example: '1999-11-11'
  /api/autorizacion/politicas:
    get:
      summary: Api para obtener el listado de politicas
      tags: ['autorizacion']
      operationId: get-api-autorizacion-politicas
      description: Api para obtener el listado de politicas
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje: 
                    type: string
                    example: 'Registro(s) obtenido(s) con exito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      filas:
                        type: array
                        items:
                          type: object
                          properties:
                            sujeto:
                              type: string
                              example: 'ADMINISTRADOR'
                            objeto:
                              type: string
                              example: '/usuarios'
                            accion:  
                              type: string
                              example: 'read'
                            app:
                              type: string
                              example: 'frontend'
                      total:
                        type: number
                        example: 1
    post:
      summary: Api para crear una nueva politica
      tags: ['autorizacion']
      operationId: post-api-autorizacion-politicas
      description: Api para crear una nueva politica
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                sujeto:
                  type: string
                  example: 'ADMINISTRADOR'
                objeto:
                  type: string
                  example: 'recurso'
                accion:
                  type: string
                  example: 'read'
                app:
                  type: string
                  example: 'frontend'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                    example: 'Registro creado con exito!'
                  finalizado:
                    type: boolean
                    example: true
                  datos:
                    type: object
                    properties:
                      sujeto:
                        type: string
                        example: 'ADMINISTRADOR'
                      objeto:
                        type: string
                        example: 'recurso'
                      accion:
                        type: string
                        example: 'read'
                      app:
                        type: string
                        example: 'frontend'
    delete:
      summary: Api para eliminar una politica
      tags: ['autorizacion']
      operationId: delete-api-autorizacion-politicas
      description: Api para eliminar una politica
      responses:
        '204':
          description: NOT_CONTENT
      parameters:
        - in: query
          name: sujeto
          schema:
            type: string
            example: ADMINISTRADOR
          description: nombre del rol (SUJETO)
        - in: query
          name: objeto
          schema:
            type: string
            example: RECURSO
          description: nombre del recurso (OBJETO)
        - in: query
          name: accion
          schema:
            type: string
            example: GET
          description: nombre de la accion (ACCION)
        - in: query
          name: app
          schema:
            type: string
            example: Backend
          description: nombre de la aplicacion
...